rename lambda folder to authentication
