+++ lambdas/authentication/package.json
