+++ lambdas/authentication/requirements.txt
@@ -1,0 +1,9 @@
+boto3==1.29.3
+botocore==1.32.3
+jmespath==1.0.1
+mccabe==0.7.0
+PyJWT==2.8.0
+python-dateutil==2.8.2
+s3transfer==0.7.0
+six==1.16.0
+urllib3==1.26.18
